Strip main function
